Relocated marked isolation engines into isolation_functions.engines
