+++ halotools/mock_observables/isolation_functions/conditional_cylindrical_isolation.py
@@ -10,9 +10,9 @@ import multiprocessing
 
 from .cylindrical_isolation import _cylindrical_isolation_process_args
 from .process_args_helpers import _conditional_isolation_process_marks
+from .engines import marked_cylindrical_isolation_engine
 
 from ..pair_counters.rectangular_mesh import RectangularDoubleMesh
-from ..pair_counters.marked_cpairs import marked_cylindrical_isolation_engine
 from ..pair_counters.mesh_helpers import _set_approximate_cell_sizes, _cell1_parallelization_indices
 
 __all__ = ('conditional_cylindrical_isolation', )
