+++ halotools/mock_observables/isolation_functions/engines/__init__.py
@@ -2,5 +2,10 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+__all__ = ('spherical_isolation_engine', 'cylindrical_isolation_engine', 
+    'marked_spherical_isolation_engine', 'marked_cylindrical_isolation_engine')
+
 from .spherical_isolation_engine import spherical_isolation_engine
 from .cylindrical_isolation_engine import cylindrical_isolation_engine
+from .marked_spherical_isolation_engine import marked_spherical_isolation_engine
+from .marked_cylindrical_isolation_engine import marked_cylindrical_isolation_engine
