+++ halotools/mock_observables/isolation_functions/engines/setup_package.py
@@ -2,7 +2,9 @@ from distutils.extension import Extension
 import os
 
 PATH_TO_PKG = os.path.relpath(os.path.dirname(__file__))
-SOURCES = ("spherical_isolation_engine.pyx","cylindrical_isolation_engine.pyx")
+SOURCES = ("spherical_isolation_engine.pyx","cylindrical_isolation_engine.pyx", 
+    "marked_spherical_isolation_engine.pyx", "marked_cylindrical_isolation_engine.pyx", 
+    "isolation_criteria_marking_functions.pyx")
 THIS_PKG_NAME = '.'.join(__name__.split('.')[:-1])
 
 def get_extensions():
