+++ halotools/mock_observables/pair_counters/marked_cpairs/__init__.py
@@ -1,11 +1,9 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-from __future__ import (absolute_import, division, print_function, unicode_literals)
+from __future__ import absolute_import, division, print_function, unicode_literals
 
 from .conditional_pairwise_distances import *
 from .marked_npairs_3d_engine import marked_npairs_3d_engine
 from .marked_npairs_xy_z_engine import marked_npairs_xy_z_engine
-from .marked_spherical_isolation_engine import marked_spherical_isolation_engine
-from .marked_cylindrical_isolation_engine import marked_cylindrical_isolation_engine
 from .velocity_marked_npairs_3d_engine import velocity_marked_npairs_3d_engine
 from .velocity_marked_npairs_xy_z_engine import velocity_marked_npairs_xy_z_engine
