+++ halotools/mock_observables/pair_counters/marked_cpairs/setup_package.py
@@ -5,7 +5,6 @@ PATH_TO_PKG = os.path.relpath(os.path.dirname(__file__))
 SOURCES = ("custom_weighting_func.pyx",
     "distances.pyx", 
     "conditional_pairwise_distances.pyx", "marked_npairs_3d_engine.pyx", 
-    "isolation_criteria_marking_functions.pyx", 
     "velocity_marked_npairs_3d_engine.pyx", 
     "velocity_marking_functions.pyx", 
     "velocity_marked_npairs_xy_z_engine.pyx")
